Update btc_hist.py
This is a very basic analysis of Bitcoin's (BTC) historical data from January 2012  to September 2020. I've tried to focus on price action and volume difference between 2 bull runs that have occured in 2013 and 2017 respectively. Even in this basic analysis, one can find useful data outputs that can be used in Economics.